feat(AssetDispose): integrate elasticsearch asset dispose;

+++ app/Services/Disposes/AssetDisposeService.php
@@ -5,17 +5,25 @@ namespace App\Services\Disposes;
 use App\DataTransferObjects\Disposes\AssetDisposeData;
 use App\Enums\Workflows\LastAction;
 use App\Enums\Workflows\Status;
+use App\Facades\Elasticsearch;
 use App\Http\Requests\Disposes\AssetDisposeRequest;
 use App\Models\Disposes\AssetDispose;
 use App\Repositories\Disposes\AssetDisposeRepository;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 
 class AssetDisposeService
 {
-    public function all()
+    public function __construct(
+        protected AssetDisposeRepository $assetDisposeRepository
+    ) {
+    }
+
+    public function all($search = null)
     {
-        return AssetDispose::query()->where('nik', auth()->user()?->nik)->get();
+        $data = Elasticsearch::setModel(AssetDispose::class)->searchMultiMatch($search, 50)->all();
+        return AssetDisposeData::collection(Arr::pluck($data, '_source'))->toCollection()->where('nik', auth()->user()?->nik);
     }
 
     public function updateOrCreate(AssetDisposeRequest $request)
@@ -23,8 +31,11 @@ class AssetDisposeService
         $data = AssetDisposeData::from(array_merge($request->all(), ['status' => Status::OPEN]))->except('employee');
         DB::transaction(function () use ($data) {
             $assetDispose = (new AssetDisposeRepository)->updateOrCreate($data);
-            $assetDispose->workflows()->delete();
+            if ($data->getKey()) {
+                $assetDispose->workflows()->delete();
+            }
             DisposeWorkflowService::setModel($assetDispose)->store();
+            $this->assetDisposeRepository->sendToElasticsearch($assetDispose, $data->getKey());
         });
     }
 
@@ -32,6 +43,7 @@ class AssetDisposeService
     {
         return DB::transaction(function () use ($assetDispose) {
             $assetDispose->workflows()->delete();
+            $this->assetDisposeRepository->deleteFromElasticsearch($assetDispose);
             return $assetDispose->delete();
         });
     }
